requirements.txt
# Web
fastapi
uvicorn[standard]
python-dotenv

# LangChain + community (tested set)
langchain==0.3.27
langchain-community==0.3.31
langchain-openai==0.2.9
langchain-text-splitters==0.3.11
langserve==0.3.0

# Embedding / OpenAI
openai>=1.0.0
tiktoken

# Vector store
faiss-cpu

# Utilities
requests
beautifulsoup4
pypdf
tqdm
httpx==0.27.2
sse-starlette

# numeric
numpy>=1.24,<1.29
